requirements.txt
asgiref==3.8.1
Django==5.1.5
django-allauth==65.3.1
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg==3.2.4
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3

africastalking~=1.2.9